Started work on dissecting STM so it passes Travis
