+++ src/main.cpp
@@ -8,8 +8,8 @@
 
 #include "wrap-hwlib.hpp"
 
-#include "enum_states.hpp"
 #include "rtc_time.hpp"
+#include "state_machine.hpp"
 #include "time_manager.hpp"
 
 int main() {
@@ -21,7 +21,7 @@ int main() {
     auto sda = hwlib::target::pin_oc(hwlib::target::pins::sda);
 
     auto clock = Time::TimeManager(scl, sda);
-    clock.setTime(0);
+    // clock.setTime(0);
 
     auto alarmsSize = clock.getAlarmArraySize();
     uint16_t alarmCounter = 0;
@@ -31,8 +31,6 @@ int main() {
 
     enum TimeManagerStates timeMngrState;
 
-    char someInput;
-
     while (true) {
         auto now = clock.getTime();
         timeMngrState = TimeManagerStates::MAIN_MENU;
@@ -40,89 +38,23 @@ int main() {
         switch (timeMngrState) {
 
         case TimeManagerStates::MAIN_MENU:
-            hwlib::cout << clock.getTime().getTotalSeconds() << hwlib::endl;
-            hwlib::cout << "Press 1 to cycle through menu, press 2 to select menu, press 3 to cancel" << hwlib::endl;
-
-            if (clock.getTime() - now >= 60) {
-                hwlib::cout << clock.getTime().getTotalSeconds() << hwlib::endl;
-                now = clock.getTime();
-            }
-
-            hwlib::cin >> someInput;
-
-            if (someInput == '1') {
-                timeMngrState = TimeManagerStates::TIMER_SELECT;
-            }
+            mainMenu(clock, timeMngrState);
             break;
 
         case TimeManagerStates::TIMER_SELECT:
-            hwlib::cout << "Timer menu? Press 2 to confirm, press 3 to cancel" << hwlib::endl;
-            hwlib::cin >> someInput;
-            if (someInput == '1') {
-                timeMngrState = TimeManagerStates::ALARM_SELECT;
-            }
-            if (someInput == '2') {
-                timeMngrState = TimeManagerStates::TIMER;
-            }
-            if (someInput == '3') {
-                timeMngrState = TimeManagerStates::MAIN_MENU;
-            }
+            timerSelect(clock, timeMngrState);
             break;
 
         case TimeManagerStates::ALARM_SELECT:
-            hwlib::cout << "Alarm menu? Press 2 to confirm, press 3 to cancel" << hwlib::endl;
-            hwlib::cin >> someInput;
-            if (someInput == '1') {
-                timeMngrState = TimeManagerStates::TIMER_SELECT;
-            }
-            if (someInput == '2') {
-                timeMngrState = TimeManagerStates::ALARM;
-            }
-            if (someInput == '3') {
-                timeMngrState = TimeManagerStates::MAIN_MENU;
-            }
+            alarmSelect(clock, timeMngrState);
             break;
 
         case TimeManagerStates::TIMER:
-            hwlib::cout << "Timer selected.Press 1 to cycle through timers, press 2 to start a timer \n Press 4 to reset a timer, "
-                           "press 5 to clear a timer \n Press 3 to return to main menu"
-                        << hwlib::endl;
-            hwlib::cin >> someInput;
-            if (someInput == '1') {
-                timerCounter++;
-            }
-            if (someInput == '2') {
-                clock.setTimer(timerCounter);
-            }
-            if (someInput == '3') {
-                timeMngrState = TimeManagerStates::MAIN_MENU;
-            }
-            if (someInput == '4') {
-                clock.resetTimer(timerCounter);
-            }
-            if (someInput == '5') {
-                clock.clearTimer(timerCounter);
-            }
+            timerMenu(clock, timeMngrState, timerCounter);
             break;
 
         case TimeManagerStates::ALARM:
-            hwlib::cout
-                << "Alarm selected.Press 1 to cycle through alarms, press 2 to start an alarms \n Press 4 to clear an alarms, "
-                   "\n Press 3 to return to main menu"
-                << hwlib::endl;
-            hwlib::cin >> someInput;
-            if (someInput == '1') {
-                alarmCounter++;
-            }
-            if (someInput == '2') {
-                timeMngrState = TimeManagerStates::SET_ALARM;
-            }
-            if (someInput == '3') {
-                timeMngrState = TimeManagerStates::MAIN_MENU;
-            }
-            if (someInput == '4') {
-                clock.clearAlarm(alarmCounter);
-            }
+            alarmMenu(clock, timeMngrState, alarmCounter);
             break;
 
         case TimeManagerStates::SET_ALARM:
